+++ packages/s3-static-assets/src/lib/constants.ts
